ifupdownmain: use run-parts for determining which hook scripts to run
Currently ifupdown2 executes scripts that are backed up by dpkg (e.g.
example.dpkg-old). This leads to potential issues with hook scripts
after upgrading. This also brings ifupdown2 on par with the behavior
of ifupdown which did not execute such scripts.

Using run-parts should be fine on any Debian installation, since it
is part of debianutils which is marked as an essential package.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>